+++ src/main/java/ru/kelcuprum/alinlib/gui/styles/AbstractStyle.java
@@ -10,6 +10,7 @@ import ru.kelcuprum.alinlib.gui.toast.ToastBuilder;
 
 import static ru.kelcuprum.alinlib.gui.Colors.BLACK_ALPHA;
 import static ru.kelcuprum.alinlib.gui.Colors.CPM_BLUE;
+import static ru.kelcuprum.alinlib.info.Player.getY;
 
 public abstract class AbstractStyle {
     public String id;
@@ -40,6 +41,24 @@ public abstract class AbstractStyle {
     public void renderBlockquoteBackground(TextBuilder builder, GuiGraphics guiGraphics, int x, int y, int width, int height, int[] colors){
         ModernStyle.renderDefaultBlockquoteBackground(builder, guiGraphics, x, y, width, height, colors);
     }
+    public int[] getCheckBoxSizes(int width, int height){
+        return new int[]{height-10, height-10};
+    }
+    public void renderCheckBox(GuiGraphics guiGraphics, int x, int y, int width, int height, boolean value){
+        int cW = getCheckBoxSizes(width, height)[0];
+        int cH = getCheckBoxSizes(width, height)[1];
+        //#if MC >= 12108
+        guiGraphics.submitOutline(x, y, cW, cH, getCheckBoxColor(value));
+        //#else
+        //$$ guiGraphics.fill(x, y, x + cW, y+1, getCheckBoxColor(value));
+        //$$ guiGraphics.fill(x, y+cH-1, x + cW, y+cH, getCheckBoxColor(value));
+        //$$ guiGraphics.fill(x, y+1, x+1, y+cH, getCheckBoxColor(value));
+        //$$ guiGraphics.fill(x+cW-1, y+1, x+cW, y+cH, getCheckBoxColor(value));
+        //#endif
+        if(value){
+            guiGraphics.fill(x+2, y+2, x+cW-2, y+cH-2, getCheckBoxColor(value));
+        }
+    }
     // -=-=-=- Разноцветные хуйни
     public int getCheckBoxColor(boolean isActive){
         return isActive ? Colors.getCheckBoxColor() : 0xFFFFFFFF;
@@ -56,6 +75,9 @@ public abstract class AbstractStyle {
 
     public int getTextColor(boolean active){
         return active ? -1 : 0xFF5F5F5F;
+    }
+    public int getCursorColor(){
+        return getEditBoxColor(true);
     }
     public int getTextTitleColor(){
         return getTextColor(true);